requirements.txt
SocksiPy_branch==1.01
termcolor==1.1.0
validators==0.20.0
